move modal stuff into /components
